Updates to fts class, plotting, forecasting, multivariate, etc.
Here's a large update with a new version of the fts class that leverages S4 objects in R. I updated fts arithmetic and plotting as well where this version of fts plotting allows for visualization of fts observed for a two-dimensional domain. In addition, I am putting FSSA forecasting for fts observed over a one-dimensional domain here. This version also allows for MFSSA to be performed on multivariate fts signals where variables might take on one or two-dimensional domains. There are many other smaller updates that are included here.